app/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
elasticsearch==7.17.9
idna==3.7
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
typing_extensions==4.12.2
urllib3==1.26.19
